Major updates
delete the frontend folder and move important files in the root of the project.
I think that was the main issue of the project lately, combine frontend/src files with the root/src folder. Import all the images from s3 buckets and finally i can see the heroImages on the project.